requirements.txt
streamlit==1.29.0
pandas==2.1.4
graphviz==0.20.1
psycopg2-binary==2.9.9
scikit-learn==1.3.2
numpy==1.26.2
